Removed exceptions folder and created one class with all of the exceptions as static nested classes
