--- src/main/java/com/google/cloudassets/discovery/exceptions/ConfigTableException.java
@@ -1,11 +1,0 @@
-package com.google.cloudassets.discovery.exceptions;
-
-/**
- * This class extends the Exception class and should be thrown when something in our configuration
- * table is not configured properly.
- */
-public class ConfigTableException extends Exception {
-    public ConfigTableException(String errorMsg, Throwable cause) {
-        super(errorMsg, cause);
-    }
-}
